requirements.txt
python-dotenv
langchain
langchain-openai
colorama
google-genai
openai